refactor(SCRUM-166): common 패키지 global로 이동
